Multi-tier estimation classes are now split into different files
